renamed and moved from BufferLayout to Pipeline
